oregon-state-scraper/agent/requirements.txt
google-adk[a2a]
google-cloud-aiplatform[agents]
requests
beautifulsoup4
